[WIP] re-write in python
This commit adds a new python based component
which uses the python awsiotsdk to subscribe
for action commands that can be used to manage
an Ubuntu Core device. This first version
supports the 'device' and 'list' actions, and
currently sends dummy responses. The component
listens on the topic 'mydevices/actions' and
sends responses on the topic 'mydevices/responses'.

This agent component is loosely based on our
reference device agent:

https://github.com/canonical/iot-agent

Signed-off-by: Tony Espy <espy@canonical.com>